// src/store/modules/projectposting/action-types.ts
export enum ActionTypes {
  SAVE_PROJECTPOSTING_STEP1 = "SAVE_PROJECTPOSTING_STEP1",
  SAVE_PROJECTPOSTING_STEP2 = "SAVE_PROJECTPOSTING_STEP2",
  SAVE_PROJECTPOSTING_STEP3 = "SAVE_PROJECTPOSTING_STEP3",
  PROJECTPOSTING = "PROJECTPOSTING",
  ADD_EMPLOYEE = "ADD_EMPLOYEE",
  EMPLOYEES = "EMPLOYEES",
}